achieve a lot card component in home page
