+++ frontend/src/app/globals.css
@@ -18,4 +18,5 @@ body {
   color: var(--foreground);
   background: var(--background);
   font-family: Arial, Helvetica, sans-serif;
+  line-height: 1.8;
 }
