+++ frontend/src/app/page.js
@@ -1,100 +1,67 @@
-import Image from "next/image";
+import AchieveALotCard from "@/components/AchieveALotCard/AchieveALotCard";
 
 export default function Home() {
-  return (
-    <div className="grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]">
-      <main className="flex flex-col gap-8 row-start-2 items-center sm:items-start">
-        <Image
-          className="dark:invert"
-          src="/next.svg"
-          alt="Next.js logo"
-          width={180}
-          height={38}
-          priority
-        />
-        <ol className="list-inside list-decimal text-sm text-center sm:text-left font-[family-name:var(--font-geist-mono)]">
-          <li className="mb-2">
-            <code className="bg-black/[.05] dark:bg-white/[.06] px-1 py-0.5 rounded font-semibold">
-              src/app/page.js
-            </code>
-            .
-          </li>
-          <li>Save and see your changes instantly.</li>
-        </ol>
+  const cardData = [
+    {
+      icon: "icon",
+      title: "Love",
+      description: "A card to express love and affection.",
+      className: "!border-b-red-500",
+    },
+    {
+      icon: "icon",
+      title: "Happiness",
+      description: "A card to bring joy and smiles.",
+      className: "!border-b-green-500",
+    },
+    {
+      icon: "icon",
+      title: "Achievement",
+      description: "Celebrate your achievements and milestones.",
+      className: "!border-b-blue-500",
+    },
+    {
+      icon: "icon",
+      title: "Nature",
+      description: "Appreciate the beauty of nature and greenery.",
+      className: "!border-b-green-500",
+    },
+    {
+      icon: "icon",
+      title: "Energy",
+      description: "Feel the warmth and energy of the sun.",
+      className: "!border-b-red-500",
+    },
+    {
+      icon: "icon",
+      title: "Night",
+      description: "Embrace the calm and quiet of the night.",
+      className: "!border-b-pink-500",
+    },
+  ];
 
-        <div className="flex gap-4 items-center flex-col sm:flex-row">
-          <a
-            className="rounded-full border border-solid border-transparent transition-colors flex items-center justify-center bg-foreground text-background gap-2 hover:bg-[#383838] dark:hover:bg-[#ccc] text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5"
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <Image
-              className="dark:invert"
-              src="/vercel.svg"
-              alt="Vercel logomark"
-              width={20}
-              height={20}
-            />
-            Deploy now
-          </a>
-          <a
-            className="rounded-full border border-solid border-black/[.08] dark:border-white/[.145] transition-colors flex items-center justify-center hover:bg-[#f2f2f2] dark:hover:bg-[#1a1a1a] hover:border-transparent text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5 sm:min-w-44"
-            href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Read our docs
-          </a>
-        </div>
-      </main>
-      <footer className="row-start-3 flex gap-6 flex-wrap items-center justify-center">
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/file.svg"
-            alt="File icon"
-            width={16}
-            height={16}
+  return (
+    <div className="bg-gray-200 p-8 flex-col flex gap-[66px] px-[141px]">
+      <div className="flex flex-col items-center">
+        <h1 className="font-extrabold text-[80px]">
+          Achieve A Lot <span className="text-red-500">With...</span>
+        </h1>
+        <h3 className="text-[20px]">
+          Sollicitudin vitae eu nulla amet ut vestibulum senectus enim. Etiam
+          cras sapien ipsum mi sit urna auctor facilisi. Mollis
+        </h3>
+      </div>
+      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 justify-center items-center bg-gray-200 gap-8">
+        {cardData.map((card, index) => (
+          <AchieveALotCard
+            key={index}
+            icon={card.icon}
+            title={card.title}
+            description={card.description}
+            className={card.className}
           />
-          Learn
-        </a>
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/window.svg"
-            alt="Window icon"
-            width={16}
-            height={16}
-          />
-          Examples
-        </a>
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://nextjs.org?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/globe.svg"
-            alt="Globe icon"
-            width={16}
-            height={16}
-          />
-          Go to nextjs.org →
-        </a>
-      </footer>
+        ))}
+      </div>
     </div>
   );
 }
